post reactions are now storing to the postreactions bridge table with the appropriate information, stashing progress to continue working on functionality
